"After attempting to checkout with a credit card and selecting a tip I get the following error: statusCode = 100 XML Validation failed Error Message: The 'isPrinted' element is invalid - The value 'false' is invalid according to its datatype 'http://www.w3.org/2001/XMLSchema:int' - The string 'false' is not a valid Int32 value. OK"
When creating a table with 16 seats, placing on hold and reopening - the order is all mixed up and makes it very difficult for the waiter to continue the order. Please place in numerical order upon reopening.


Tables Floor Map Builder in BO shows layouts by location, but the device who is assigned to a location and has "Download Location Inventory only:" set to YES can see all tables in all locations. If the layout has them placed in the same area, they show up in the device as one on top of another.